move intermediates manager into core
